DEV-684: Fix issue for filtering gene amplifications in patient report

+++ hmf-common/src/main/java/com/hartwig/hmftools/common/copynumber/CopyNumber.java
@@ -4,7 +4,7 @@ import com.hartwig.hmftools.common.region.GenomeRegion;
 
 public interface CopyNumber extends GenomeRegion {
 
-    int value();
+    double value();
 
     default CopyNumberAlteration alteration() {
         return CopyNumberAlteration.fromCopyNumber(value());
